% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunConfetti.R
\name{confetti.run}
\alias{confetti.run}
\title{Run the CONFETTI model}
\usage{
confetti.run(pars = c(metaSR = 300, metaCV = 1, m = 0.1, rmax = 10, aRec =
  0.005, disp = 30, CNDD.m = 1, CNDD.cv = 0), nRep = 1, nGen = 100,
  nSteps.out = 1, nTrees = 10000, Xext = 500, Yext = 500,
  dist.max = 100, avg = FALSE)
}
\arguments{
\item{pars}{A vector of eight numbers that include the model parameters for CONFETTI. The vector can be named but need not. Note that the order of parameters matters.

\enumerate{
   \item \bold{metaSR} - Species richness of the metacommunity
   \item \bold{metaCV} - Coefficient of variation (sd/mean) of the metacommunity lognormal abundance distribution
\item \bold{m} - Immigration rate (Interval [0,1])
\item \bold{rmax} - Radius of interactions neighborhood (Interval [0,20])
\item \bold{aRec} - Parameter for recruitment competition (Interval [0,0.1])
\item \bold{disp} - Mean dispersal distance
\item \bold{CNDD.m} - Mean conspecific negative density dependence (CNDD)
\item \bold{CNDD.cv} - Coefficient of variation (sd/mean) of CNDD among species
}}

\item{nRep}{Number of replicate runs with the same parameter set}

\item{nGen}{Number of generations simulated. This means the model is simulates nGen*nTrees birth-death events}

\item{nSteps.out}{Number of time steps with model output. If nSteps.out == 1 there is only model output from the last simulated time step. If nSteps.out > 1 there is output over time. The output time itnerval is calculated as nGen/nSteps.out. The initial condition is included in the output. Temporal output works only on combination with avg = F.}

\item{nTrees}{Number of trees in the local community}

\item{Xext}{Size of the local community in x-direction}

\item{Yext}{Size of the local community in y-direction}

\item{dist.max}{Maximum distance (in meters) over which the spatial patterns F(r) and PCF(r) are calculated. The step size is always 1 meter}

\item{avg}{logical variable: Should the non-temporal model output be averaged over the replicate runs? This works only if nSteos.out == 1.}
}
\value{
The model simulates virtual tree censuses with coordinates and species identities for each tree. The census data of the last simulated time step is only provided as output when nRep = 1 and avg = F. In this case the output is a list with four elements:
\enumerate{
   \item \bold{census} - a dataframe with x,y coordinates and species ID for every tree in the simulated community
   \item \bold{abundance} - a vector with the number of individuals for every species
   \item \bold{species} - a dataframe with species properties: (1) the species ID, (2) the relative abundance in the metacommunity, (3) the mean dispersal distance, (4) the species-specifc conspecific negative density dependence (CNDD)
   \item \bold{summary.stats} - a list with eigth elements that includes non-spatial and spatial summary statistics of the simulated community. When nSteps.out == 1 the summary statistics are provided as scalars and vectors and when nStpes.out > 1 as vectors and matrices. The list elements are:
\enumerate{
   \item \bold{nSpecies} - Number of species
   \item \bold{Shannon} - Shannon diversity index \eqn{H = - \sum log(pi) * pi}
   \item \bold{Simpson} - Simpson diversity index \eqn{S = 1 - \sum 1/pi^2}, where pi is the relative abundance of species i. This essentially estimates the probability of randomly drawing two individuals from different species (probability of interspecific encounter)
   \item \bold{SAD} - species abundance distribution with logarithmic abundance classes (1, 2-3, 4-7, 8-15, ..., >=2048 individuals)
   \item \bold{Area} - sampling areas for species area relationship in m^2
   \item \bold{SAR} - species-area relationship: average species numbers in quadrats of different sizes
   \item \bold{Fr} - Proportion of conspecific neighbors: This estimates the probability that two trees seperated by distance r belong to the same species.
   \item \bold{PCF} - pair-correlation function. Measure of spatial aggregation of regularity across scales. A value of 1 indicates a random pattern, PCF > 1 indicates aggregation and PCF < 1 regularity (=repulsion of individuals)
}
}
When nRep > 1 or avg = T only the list with the summary statistics is returned.
}
\description{
Run the CONFETTI model
}
\examples{
#Run CONFETTI with standard parameters
confetti.run()

# Define your own parameter vector
parvec <- c(500, 2.0, 0.01, 20, 0.005, 55, 2.0, 0.2)
out1 <- confetti.run(pars = parvec)

# plot model output
plot(0:11, out1$SAD,type="b",xlab="log2(Abundance)", ylab="No. of species")
plot(out1$Area, out1$SAR, type="b", log="xy", xlab="Area [m2]", ylab="No. of species")

}
\references{
May, F.; Huth, A. & Wiegand, T. (2015) Moving beyond abundance distributions: neutral theory and spatial patterns in a tropical forest Proceedings of the Royal Society of London B: Biological Sciences, 282, 20141657

May, F.; Wiegand, T.; Lehmann, S. & Huth, A. (2016) Do abundance distributions and species aggregation correctly predict macroecological biodiversity patterns in tropical forests? Global Ecology and Biogeography, 25, 575-585
}

