% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunConfetti.R
\name{confetti.run}
\alias{confetti.run}
\title{Run the CONFETTI model}
\usage{
confetti.run(pars = c(metaSR = 100, metaCV = 1, m = 0.1, rmax = 10, aRec =
  0.005, disp.m = 30, disp.cv = 0, CNDD.m = 1, CNDD.cv = 0, pRec.m = 1, pRec.cv
  = 0, trade1.CNDD.pRec = 0, a.CNDD.pRec = 0, b.CNDD.pRec = 0, c.CNDD.pRec = 0,
  trade2.CNDD.abund = 0, a.CNDD.abund = 0, b.CNDD.abund = 0, c.CNDD.abund = 0,
  trade3.disp.pRec = 0, a.disp.pRec = 0, b.disp.pRec = 0, c.disp.pRec = 0),
  nRep = 1, nGen = 100, nSteps.out = 1, nTrees = 10000, Xext = 500,
  Yext = 500, dist.max = 100, bin.width = 1, meta.SAD = 1,
  avg = FALSE)
}
\arguments{
\item{pars}{Model parameters of CONFETTI. Either a vector of length 23 or a named vector of variable length. The names have to match the following list. When the parameter is unnamed the order has to correspond to the following list.

\enumerate{
   \item \bold{metaSR} - Species richness of the metacommunity
   \item \bold{metaCV} - Coefficient of variation (sd/mean) of the metacommunity lognormal abundance distribution
\item \bold{m} - Immigration rate (Interval [0,1])
\item \bold{rmax} - Radius of interactions neighborhood (Interval [0,20])
\item \bold{aRec} - Parameter for recruitment competition (Interval [0,0.1])
\item \bold{disp.m} - Mean dispersal distance of species
\item \bold{disp.cv} - Coefficient of variation (sd/mean) of dispersal distance among species
\item \bold{CNDD.m} - Mean conspecific negative density dependence (CNDD)
\item \bold{CNDD.cv} - Coefficient of variation (sd/mean) of CNDD among species
\item \bold{pRec.m} - mean recruitment probability without competition in interval [0.001; 1]
\item \bold{pRec.cv} - Coefficient of variation (sd/mean) of recruitment probability among species
\item \bold{trade1.CNDD.pRec} - mode for trade-off between CNDD and recruitment probability. There are three different possible modes (parameterizations) for each of the three trade-offs:
\itemize {
   \item 0: no trade-off
   \item 1: linear: \eqn{y = a + b * x}
   \item 2: logarithmic: \eqn{y = a + b * log(x)}
   \item 3: exponential: \eqn{y = a + b * exp(c * x)}
}
\item \bold{a.CNDD.pRec} - parameters for CNDD-recruitment trade-off
\item \bold{b.CNDD.pRec}
\item \bold{c.CNDD.pRec}
\item \bold{trade2.CNDD.abund} - mode for trade-off between CNDD and metacommunity relative abundance:
\item \bold{a.CNDD.abund} - parameters for CNDD-metacommunity abundance trade-off
\item \bold{b.CNDD.abund}
\item \bold{c.CNDD.abund}
\item \bold{trade3.disp.pRec} - mode for trade-off between dispersal and recruitment probability:
\item \bold{a.disp.pRec} - parameters for dispersal-recruitment trade-off
\item \bold{b.disp.pRec}
\item \bold{c.disp.pRec}
}}

\item{nRep}{Number of replicate runs with the same parameter set}

\item{nGen}{Number of generations simulated. This means the model is simulates nGen*nTrees birth-death events}

\item{nSteps.out}{Number of time steps with model output. If nSteps.out == 1 there is only model output from the last simulated time step. If nSteps.out > 1 there is output over time. The output time interval is calculated as nGen/nSteps.out. The initial condition is included in the output. Temporal output works only on combination with avg = F.}

\item{nTrees}{Number of trees in the local community}

\item{Xext}{Size of the local community in x-direction}

\item{Yext}{Size of the local community in y-direction}

\item{dist.max}{Maximum distance (in meters) over which the spatial patterns F(r) and PCF(r) are calculated.}

\item{bin.width}{The distance bin-width used for the calculation of the spatial patterns F(r) and PCF(r),}

\item{meta.SAD}{The abundance distribution of the metacommunity: (0) uniform distribution, (1) log-normal distribution with coefficient of variation metaCV.}

\item{avg}{Logical variable: Should the non-temporal model output be averaged over the replicate runs? This works only if nSteps.out == 1.}
}
\description{
Run the CONFETTI model
}
\details{
There are some constraints on parameter values that override the parameter settings. Species traits, i.e. recruitment probability, dispersal distance, and CNDD are simulated from probability distributions. To avoid biologically unrealistiv values a few constraints are implemented: Recruitment probability is restricted to the interval [0.001; 1]; For dispersal distance there is a lower bound of 0.1 m, and for CNDD the lower bound is 1, which means intraspecific competition equals interspedific competition. This lower bound excludes conspecific positive density dependence.

There can be only one trade-off that involves CNDD. If the parameter settings imply a trade-off between CNDD and recruitment as well as between CNDD and metacommunity abundance, the second one is ignored and just the trade-off with recruitment probability is simulated.
}
\examples{
#Run CONFETTI with standard parameters
confetti.run()

# Define your own parameter vector
parvec <- c(500, 2.0, 0.01, 20, 0.005, 55, 0.2, 2.0, 0.2, 1.0, 0.0,
            0, 1.0, 0.0, 0.0, 0, 1.0, 0.0, 0, 1.0, 0.0, 0.0)
out1 <- confetti.run(pars = parvec)

# plot model output
plot(0:11, out1$SAD,type="b",xlab="log2(Abundance)", ylab="No. of species")
plot(out1$Area, out1$SAR, type="b", log="xy", xlab="Area [m2]", ylab="No. of species")

}
\references{
May, F.; Huth, A. & Wiegand, T. (2015) Moving beyond abundance distributions: neutral theory and spatial patterns in a tropical forest Proceedings of the Royal Society of London B: Biological Sciences, 282, 20141657

May, F.; Wiegand, T.; Lehmann, S. & Huth, A. (2016) Do abundance distributions and species aggregation correctly predict macroecological biodiversity patterns in tropical forests? Global Ecology and Biogeography, 25, 575-585
}

